PyQt Settings GUI for SCCwatcher
Working on revamping this whole damn thing. SCC.ini is a massive
embarrassment that I've kept around from the original script I took over
and its time to go. Its going to become easier to use via the scc.ini as
well as an external (as well as optional) PyQt-powered GUI.