Update to the settings manager
Been sitting on this one for a while because I was sure there was a
better way to do this. It was either the smaller but uglier solution or
the much larger but obvious solution. I basic way around this basically
doubled the amount of code needed and it wouldn't be reused at all so I
opted to make this piece as small as possible. It might be possible to
further refine this in the future.

This is still a work in progress. It has yet to convert the ui elements
into options items but thats to come.