Updated fonts/layout for compatibility
Previous layout used MS Shell Dlg 2 for most fonts. This font turns out
to be a virtual font that points to Tahoma on my own PC. Unfortunately
Tahoma isn't a font that is cross-platform compatible (even if you can
install it, it doesn't look good at all). I've decided to use the old,
tried-and-true Ariel font for everything. There are only minor
differences in appearance after switching and now it looks the same
under windows and linux. Ariel ftw.

Forgot a file in the last sync, mistook one file for another.