More UI Actions
Added more UI actions. These ones were pretty significant. I'm tackling
the hardest problems first, after that it should be an easy thing to get
the rest done. The hardest part is managing the data associated with the
watch and avoid lists and I feel like I'm not doing a very good job
keeping the code clean and consistent.