Loading is done
Load system is now complete. There are still some nice things to get
done like a file selection window for loading but the code needed to
load the options file and update the GUI with the data is finished.
There are some weird querks of PyQt I can't work around but I doubt
anyone will notice. This is my last update before I go on vacation
tomorrow, hopefully the updates will be more frequent over the next
month or so.