More logic updates
Almost done, down to the last few items before the GUI is officially
finished. This update finished off the Save-as dialog, code to detect
when save-as is needed, code to detect when the file the user tried to
load is not a valid sccwatcher 2.0 settings file, and code to stop
"double loads" so only the current file is loaded. If the user tries to
load a file while another is loaded it will prompt the user to ask if
they want to save or not.