small update
Just a small update. I got the change detection system in place so it
warns the user when they make a change to their current file. I also
implemented the New action as well as detection of a new file which
turns the save dialog into a save-as dialog (before it would error out).
I am slowly checking things off the list.