Feature freeze
As of now, the GUI is finally in a feature freeze. I am finished adding
all new features and am just tidying things up for the integration into
the actual script. Will be uploading some test binaries soon as well.
This update specifically contains code for the "check for updates"
button as well as fixes a bug in the Save-as dialog when launched
through the normal save action when there isn't already a file set to
save to. When the user canceled saving before it would change the window
title from "New Settings File" to "()". Now it stays with "New Settings
File" until the user actually chooses a file to save to.