First beta build is up
I have uploaded the first beta .exe file to test if anyone wants it.
I've uploaded the spec file in-case you want to try and built it
yourself. Make sure you have upx in the build folder otherwise it won't
build. Changes in this build include: Change to the behavior of size
limit entry boxes. Now when the lower or upper limit is out of bounds,
it will change the entry boxes' background color to indicate the error.
Updated warning message boxes with icons and more exact error messages
and text placement. Minor bug fixes.