finally getting on
Been thinking about this one for a while and I finally decided to go
with the multiprocessing library for communication between the GUI and
the script. This does unfortunately mean I have to manage the threads to
make sure it doesn't block the xchat/hexchat GUI but its not a big
issue. It should all work.

This is also one of the first builds that officially breaks the script.
This script will NOT work if you try to load it up in xchat/hexchat. No
idea when this will be working tbh.