Semi-working gui/script communication
So the autodl part of this might not work (it shouldn't, but I saw it
grabbing stuff earlier lol) but the communication between the beta5 GUI
and this script is working at a basic level (keep an eye on the
console). Unfortunately it seems to cause hexchat to crash on load in
linux and in windows it always spits out a fatal error but keeps on
working. But it crashes when you unload the script.

Since I have used threads successfully and prolifically in xchat/hexchat
in the past I expect to be able to get this to work eventually. I will
readjust the way I am creating/managing these threads and try again. If
all else fails the backup is a sufficiently dirty system of reoccurring
timers calling functions that check a file on the computer (bleh I
know). Not elegant in any way whatsoever but whatever its works in the
end.