Done trying to make this work
Ok I give up on the idea of having a nice, mutithreaded interprocess
communication system that actually works with xchat/hexchat's horrible
handling of threads. I already had a working system but noooo, hexchat
dies at the sight of it. Im convinced the issue has to do with the
expanding call stack inside the threads due to the use of callback
functions. Rearranging the system so that each part was independent of
another was proving to be difficult to accomplish. Every time I take a
break and come back I'm absolutely lost as to where I should begin. I
finally realized this was just ridiculous and so I'm done working
further on this dead-end idea.

The reason i'm committing this soon-to-be-deleted code is just in-case
my future endeavors fail I can come back to this and maybe try again.
Also I couldn't just delete all this work, I really thought through a
lot of the design. It would work eventually but its just too difficult
to be the right way about this.