Minor bug fix update
Just tcb, getting rid of terrible newbie bugs. Some checks to make sure
a good sccwatcher v2 ini is being loaded and not an old ini. This also
has some default settings built into it now, like dupechecking and SSL
downloading being on by default. The defaults in here will be in-line
with the defaults the GUI uses when creating a new file by release. Its
on my todo list :).