Script - GUI communication working
Finally got the script to GUI communication working, unfortunately its a
really dirty way to do it (reading from a file). I wanted to use sockets
to allow remote control of the script but hexchat is a massive bitch
about threads so this is as good as it gets unfortunately.

Some other fixes are in here too but tbh it has been months since I
touched the code so I have no idea whats even been updated. Looks like
minor fixes here and there. Lots of compatibility changes to work with
the new data format introduced with the new settings file. One change
for sure was the addition of watch-specific options which override the
global options. Looks like its mostly just updates to make the script
compatible with all the new features v2 offers.