Broken external_command fixed
It just wouldn't be a TRB release without a missed bug. So pr0. Weird
thing is, I tested the damn thing so many times and it always worked,
then the one time I test it after committing it fails. El oh els. Went
ahead and changed some error/verbose messages while I was at it so its
not all bad.