Non-debug socket coms
Everything works great except if you reload the script while the GUI is
open, hexchat crashes. It has to do with closing down the sockets
properly but as far as I can tell I am doing it properly. I might have
to send a shutdown signal to the GUI and have it close from its side
before closing the script socket. Too tired to mess with it now.