Crash fix attempt #1
There is this nasty crash that happens seemingly at random and is
totally unpredictable thus far. I've had it crash after a few minutes,
and I've let it run for days with no crash. The only thing in common is
the crash happens when the GUI sends a reload scc.ini command. Since
this is a socket-specific error I decided to rewrite the whole server
using select in an attempt to mitigate any issue. This is the only
problem holding up v2.1. I also fixed many print statements not being
verbose() commands. Now all information outputs to the user-specified
tab or current tab.